port(phase2): armour is one byte, and its adds wrap like the original's
`20ae:38b2` is a single byte and every write to it is a byte op:

    1000:bfa7  inc [0x38b2]              the abibas suit
    1000:c107  add byte [0x38b2],0x2     the jacket
    1000:c1b1  inc [0x38b2]              the adidas suit's delta
    1000:c2f8  add byte [0x38b2],0x2     the crutaya's delta
    1000:e8d6  inc [0x38b2]              the gym's abs arm

`combat_trace.json` says the same thing about the enemy's copy --
`e_armor_3968` is recorded with `"width": 1` -- and `src/save.rs` already
declared `armour: u8`. `Fighter::armor` was `u16`.

Narrowed to `u8`, and the widening's fingerprints come out with it:

* `src/combat.rs`'s damage step carried `(defender.armor & 0x00ff) as i16`.
  That mask existed only to recover the byte from a wider field; it is
  `i16::from(..)` now.
* `Game::trained_armour` opened with `self.player.armor as u8` for the same
  reason -- the recompute at `1000:e3a4` reads a byte. The cast is gone and
  its `wrapping_sub` chain is now plain byte arithmetic, which is what it
  was modelling.
* `persist` converts in neither direction any more.

**The five adds now wrap.** `inc` and `add byte` wrap at 255; `+=` on a `u8`
panics there in a debug build, so a faithful port needs `wrapping_add` and
not an implicit widening. This is the same trap the trained-armour recompute
already documented from the other side -- armour 1 with the crutaya kozhanka
is 253, not -3 -- and narrowing the field is what makes both ends of it
consistent.

Test fixtures read `20ae:38b2` and `20ae:3968` as the bytes they are rather
than widening them, and `tests/combat_vectors.rs`'s `FighterSpec` declares
`armor: u8` (its 590 captured values run 0..80, well inside a byte).

`cargo test` 460 passed, 0 failed. `difftest` 363/363. clippy and fmt clean.

+++ src/combat_dispatch.rs
@@ -318,7 +318,7 @@ pub fn backup_round(
     rng: &mut Rng,
     backup: &mut Backup,
     district: u8,
-    enemy_armor: u16,
+    enemy_armor: u8,
     enemy_hp: i32,
     cred: &mut i32,
 ) -> Option<Fought> {
@@ -538,7 +538,7 @@ mod tests {
         };
         // 0..=8 covers all three remainders twice over; `div` truncates, so
         // armour 2 costs nothing and armour 3 costs one.
-        for armor in 0u16..=8 {
+        for armor in 0u8..=8 {
             let mut rng = Rng::new(11);
             let mut b = backup_at(3);
             let mut cred = 500;
